requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-dotenv==0.10.5
python-http-client==3.2.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.1
six==1.14.0
sqlparse==0.3.0
twilio==6.35.4
urllib3==1.25.8
whitenoise==5.0.1
